+++ CHANGELOG.md
@@ -4,6 +4,13 @@ The format of this changelog is based on
 [Keep a Changelog](https://keepachangelog.com/), and this project adheres to
 [Semantic Versioning](https://semver.org/).
 
+## Unreleased
+
+  - Added `WithDirection <: GeometryEntityStyle` to annotate geometry entities with a direction (CCW from +x in local frame). The direction transforms with the entity under rotations and reflections, allowing extraction of the final global direction for use in simulation configuration.
+  - Added `SolidModels.check_port_connectivity`, using `SolidModels.connected_components` to report ports as `:open`, `:short`, `:floating`, or `:missing`
+  - Added `detect_non_boundary_contacts=false` keyword argument to `SolidModels.connected_components`; when `true`, 1d edges embedded in the interior of 2D surfaces (like the feet of staple air bridges) will be treated as connecting
+  - Added `examples/DemoQPU17/solidmodel.jl` demonstrating large-scale SolidModel construction and configuration, including a check for open charge lines and shorted flux lines with the new functionality above
+
 ## 1.15.0 (2026-06-14)
 
   - Added `SolidModels.populate_size_fields!(cs::AbstractCoordinateSystem)`